requirements.txt
streamlit==1.30.0
gspread==5.7.2
google-auth==2.22.0